[Refactor] 메인 SRP 적용 (#9)
