--- Tving_CloneProject/Tving_CloneProject/Scene/Main/Views/MainPosterView.swift
@@ -1,27 +1,0 @@
-////
-////  MainPosterView.swift
-////  Tving_CloneProject
-////
-////  Created by 윤희슬 on 4/22/24.
-////
-//
-//import UIKit
-//
-//class MainPosterView: UIView {
-//    
-//    // MARK: - UI Properties
-//    
-//    // MARK: - Properties
-//    
-//    override init(frame: CGRect) {
-//        super.init(frame: frame)
-//    }
-//    
-//    required init?(coder: NSCoder) {
-//        fatalError("init(coder:) has not been implemented")
-//    }
-//    
-//}
-//
-//
-//// MARK: - Private Methods
