+++ Tving_CloneProject/Tving_CloneProject/Scene/Main/Views/MainView.swift
@@ -1,0 +1,180 @@
+//
+//  MainView.swift
+//  Tving_CloneProject
+//
+//  Created by 윤희슬 on 6/3/24.
+//
+
+import UIKit
+
+import SnapKit
+import Then
+
+final class MainView: UIView {
+    
+    // MARK: - UI Properties
+
+    lazy var mainCollectionView = UICollectionView(frame: .zero, collectionViewLayout: self.makeFlowLayout())
+
+    
+    // MARK: - Properties
+    
+    private var mainViewModel: MainViewModel = MainViewModel()
+    
+    private var tabBarHeight: CGFloat = 0
+    
+    
+    // MARK: - Life Cycle
+    
+    init(tabBarHeight: CGFloat) {
+        self.tabBarHeight = tabBarHeight
+        super.init(frame: .zero)
+        
+        setHierarchy()
+        setLayout()
+        setStyle()
+    }
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+
+}
+
+private extension MainView {
+    
+    func setHierarchy() {
+        self.addSubview(mainCollectionView)
+    }
+    
+    func setLayout() {
+        mainCollectionView.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
+    }
+    
+    func setStyle() {
+        mainCollectionView.do {
+            $0.backgroundColor = UIColor(resource: .black)
+            $0.showsVerticalScrollIndicator = false
+            $0.contentInsetAdjustmentBehavior = .never
+        }
+    }
+    
+    func makeFlowLayout() -> UICollectionViewCompositionalLayout {
+        
+        return UICollectionViewCompositionalLayout { section, ev -> NSCollectionLayoutSection? in
+            
+            switch self.mainViewModel.dataSource[section] {
+            case .mainPoster:
+                return self.makeMainPosterLayout()
+            case .recommendedContents, .paramounts:
+                return self.makeImageNTitleLayout()
+            case .popularLiveChannel:
+                return self.makePopularLiveChannelLayout()
+            case .categories:
+                return self.makeImageOnlyLayout()
+            }
+        }
+    }
+    
+    func makeMainPosterLayout() -> NSCollectionLayoutSection {
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalHeight(1))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        
+        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalHeight(550 / 812))
+        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
+        
+        let section = NSCollectionLayoutSection(group: group)
+        section.orthogonalScrollingBehavior = .groupPaging
+        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 23, trailing: 0)
+        
+        let footer = makePageControlButtonView()
+        section.boundarySupplementaryItems = [footer]
+        
+        return section
+    }
+    
+    func makeImageNTitleLayout() -> NSCollectionLayoutSection {
+        
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalHeight(1))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        item.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 8)
+        
+        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(98 / 375), heightDimension: .fractionalHeight(170 / 812))
+        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
+        group.contentInsets = NSDirectionalEdgeInsets(top: 14, leading: 0, bottom: 0, trailing: 0)
+        
+        let section = NSCollectionLayoutSection(group: group)
+        section.orthogonalScrollingBehavior = .continuous
+        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 15, bottom: 40, trailing: 0)
+        
+        let header = makeHeaderView()
+        section.boundarySupplementaryItems = [header]
+        
+        return section
+    }
+    
+    func makePopularLiveChannelLayout() -> NSCollectionLayoutSection {
+        
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalHeight(1))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        item.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 8)
+        
+        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(160 / 375), heightDimension: .fractionalHeight(140 / 812))
+        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
+        group.contentInsets = NSDirectionalEdgeInsets(top: 8, leading: 0, bottom: 0, trailing: 0)
+        
+        let section = NSCollectionLayoutSection(group: group)
+        section.orthogonalScrollingBehavior = .continuous
+        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 15, bottom: 40, trailing: 0)
+        
+        let header = makeHeaderView()
+        section.boundarySupplementaryItems = [header]
+        
+        return section
+    }
+    
+    func makeImageOnlyLayout() -> NSCollectionLayoutSection {
+        
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalHeight(1))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        item.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 8)
+        
+        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(0.3), heightDimension: .fractionalHeight(58 / 812))
+        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
+        
+        let section = NSCollectionLayoutSection(group: group)
+        section.orthogonalScrollingBehavior = .groupPaging
+        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 15, bottom: tabBarHeight + 10, trailing: 0)
+        
+        return section
+    }
+    
+    func makeHeaderView() -> NSCollectionLayoutBoundarySupplementaryItem {
+        
+        let headerSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(25))
+        let header = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: headerSize,
+                                                                 elementKind: BasicHeaderView.elementKinds,
+                                                                 alignment: .top)
+        return header
+        
+    }
+    
+    func makePageControlButtonView() -> NSCollectionLayoutBoundarySupplementaryItem {
+        
+        let footerSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(36))
+        let footer = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: footerSize,
+                                                                 elementKind: PageControlButtonView.elementKinds,
+                                                                 alignment: .bottom)
+        return footer
+        
+    }
+    
+
+}
